LangChain/Chat-Multiple-PDFs/requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
faiss-cpu
openai
huggingface-hub
InstructorEmbedding
sentence-transformers
langchain_community